Renamed segmentRequest to mediaSegmentRequest to align with the naming convention in the hls spec
